Make `utterance_id` 64-bit.

+++ drivers/apple_embedded/tts_apple_embedded.h
@@ -44,7 +44,7 @@
 
 @interface GDTTTS : NSObject <AVSpeechSynthesizerDelegate> {
 	bool speaking;
-	HashMap<id, int> ids;
+	HashMap<id, int64_t> ids;
 
 	AVSpeechSynthesizer *av_synth;
 	List<DisplayServer::TTSUtterance> queue;
@@ -55,6 +55,6 @@
 - (void)stopSpeaking;
 - (bool)isSpeaking;
 - (bool)isPaused;
-- (void)speak:(const String &)text voice:(const String &)voice volume:(int)volume pitch:(float)pitch rate:(float)rate utterance_id:(int)utterance_id interrupt:(bool)interrupt;
+- (void)speak:(const String &)text voice:(const String &)voice volume:(int)volume pitch:(float)pitch rate:(float)rate utterance_id:(int64_t)utterance_id interrupt:(bool)interrupt;
 - (Array)getVoices;
 @end
